fixed ffmpeg runtime error on windows

+++ src/ofxFFmpegRecorder.cpp
@@ -389,8 +389,8 @@ bool ofxFFmpegRecorder::startCustomRecord()
     // IN
     args.push_back("-y");
     args.push_back("-an");
-    args.push_back("-r " + std::to_string(m_Fps));
-    args.push_back("-framerate " + std::to_string(m_Fps));
+    args.push_back("-r " + std::to_string(static_cast<int>(m_Fps)));
+    args.push_back("-framerate " + std::to_string(static_cast<int>(m_Fps)));
     args.push_back("-s:v " + std::to_string(static_cast<unsigned int>(m_VideoSize.x)) + "x" + std::to_string(static_cast<unsigned int>(m_VideoSize.y)));
     args.push_back("-f rawvideo");
     args.push_back("-pix_fmt " + mInputPixFmt);
@@ -400,8 +400,8 @@ bool ofxFFmpegRecorder::startCustomRecord()
     // OUT
     args.push_back("-c:v " + m_VideCodec);
     args.push_back("-b:v " + std::to_string(m_BitRate) + "k");
-    args.push_back("-r " + std::to_string(m_Fps));
-    args.push_back("-framerate " + std::to_string(m_Fps));
+    args.push_back("-r " + std::to_string(static_cast<int>(m_Fps)));
+    args.push_back("-framerate " + std::to_string(static_cast<int>(m_Fps)));
     args.push_back("-pix_fmt " + mOutputPixFmt );
 
     std::copy(m_AdditionalOutputArguments.begin(), m_AdditionalOutputArguments.end(), std::back_inserter(args));
